feat: add `QuantizePerToken` and `RmsnormQuantization` for plugin node specialization
