+++ src/ditto/fx/nodes/__init__.py
@@ -19,7 +19,14 @@ from .get_attr import GetAttr
 from .node_specialization import NodeSpecialization
 from .operator import GetItem
 from .placeholder import Placeholder
-from .plugins import Fp8RowwiseGemm, Gemm, GPTAttention, WeightOnlyGroupwiseQuantMatmul, WeightOnlyQuantMatmul
-from .rmsnorm_quantization import RmsnormQuantization
+from .plugins import (
+    Fp8RowwiseGemm,
+    Gemm,
+    GPTAttention,
+    QuantizePerToken,
+    RmsnormQuantization,
+    WeightOnlyGroupwiseQuantMatmul,
+    WeightOnlyQuantMatmul,
+)
 from .rope import Rope
 from .scaled_dot_product_attention import ScaledDotProductAttention
