+++ src/ditto/fx/nodes/plugins.py
@@ -21,6 +21,8 @@ from ..targets import (
     Fp8RowwiseGemmPlugin,
     GemmPlugin,
     GPTAttentionPlugin,
+    QuantizePerTokenPlugin,
+    RmsnormQuantizationPlugin,
     WeightOnlyGroupwiseQuantMatmulPlugin,
     WeightOnlyQuantMatmulPlugin,
 )
@@ -104,6 +106,62 @@ class GPTAttention(CallFunction):
     @classmethod
     def validate_node(cls, node: Node) -> bool:
         return isinstance(node.target, GPTAttentionPlugin)
+
+
+class QuantizePerToken(CallFunction):
+    """A plugin specialization representing a quantize per token plugin node.
+
+    Attributes:
+        this (Node): The first input node
+        clamp_val (Node): The clamp value node
+    """
+
+    this: Node
+    clamp_val: Node
+
+    @property
+    def target(self) -> QuantizePerTokenPlugin:
+        assert isinstance(t := super().target, QuantizePerTokenPlugin)
+        return t
+
+    @classmethod
+    def possible_targets(cls) -> tuple[Callable[..., Any], ...]:
+        return ()
+
+    @classmethod
+    def validate_node(cls, node: Node) -> bool:
+        return isinstance(node.target, QuantizePerTokenPlugin)
+
+
+class RmsnormQuantization(CallFunction):
+    """A plugin specialization representing a rmsnorm quantization plugin node.
+
+    Attributes:
+        this (Node): The first input node
+        weight (Node): The second input node (expected to be a weight tensor)
+        bias (Node): The third input node (expected to be a bias tensor)
+        scale (Node): The fourth input node (expected to be a scale tensor)
+        clamp_val (Node | None): The fifth input node (expected to be a clamp value tensor)
+    """
+
+    this: Node
+    weight: Node
+    bias: Node
+    scale: Node
+    clamp_val: Node | None = None
+
+    @property
+    def target(self) -> RmsnormQuantizationPlugin:
+        assert isinstance(t := super().target, RmsnormQuantizationPlugin)
+        return t
+
+    @classmethod
+    def possible_targets(cls) -> tuple[Callable[..., Any], ...]:
+        return ()
+
+    @classmethod
+    def validate_node(cls, node: Node) -> bool:
+        return isinstance(node.target, RmsnormQuantizationPlugin)
 
 
 class WeightOnlyQuantMatmul(CallFunction):
